+++ libnetwork/ipams/drivers.go
@@ -1,4 +1,4 @@
-package builtin
+package ipams
 
 import (
 	"github.com/docker/docker/libnetwork/ipamapi"
